+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - filter: fix memcmp decode data in the config ([#118](https://github.com/lamports-dev/richat/pull/118))
 - richat: do not send Block message to connected richat ([#122](https://github.com/lamports-dev/richat/pull/122))
 - plugin: disable accounts on snapshot loading ([#124](https://github.com/lamports-dev/richat/pull/124))
+- richat: use flag instead of counter in client state ([#125](https://github.com/lamports-dev/richat/pull/125))
 
 ### Features